requirements.txt
txtai
atlassian-python-api
python-dotenv
